+++ straxen/contexts.py
@@ -1,6 +1,8 @@
 from immutabledict import immutabledict
 import strax
 import straxen
+import numpy as np
+
 
 common_opts = dict(
     register_all=[
@@ -26,7 +28,7 @@ xnt_common_config = dict(
     n_nveto_pmts=120,
     n_tpc_pmts=straxen.n_tpc_pmts,
     n_top_pmts=straxen.n_top_pmts,
-    gain_model=('to_pe_constant', 'gain_2e6HVmap_cutoff_1310'),
+    gain_model=("CMT_model", ("to_pe_model", "ONLINE")),
     channel_map=immutabledict(
         # (Minimum channel, maximum channel)
         # Channels must be listed in a ascending order!
@@ -40,18 +42,23 @@ xnt_common_config = dict(
         nveto=(2000, 2119),
         nveto_blank=(2999, 2999)),
     nn_architecture=straxen.aux_repo + 'f0df03e1f45b5bdd9be364c5caefdaf3c74e044e/fax_files/mlp_model.json',
-    nn_weights=straxen.aux_repo + 'f0df03e1f45b5bdd9be364c5caefdaf3c74e044e/fax_files/mlp_model.h5', )
+    nn_weights=straxen.aux_repo + 'f0df03e1f45b5bdd9be364c5caefdaf3c74e044e/fax_files/mlp_model.h5', 
+    # Clustering/classification parameters
+    s1_max_rise_time=100,
+)
 
-# Plugins in these files are nT only (NB: pulse&peak(let) processing are
-# registered for High Energy plugins.)
-xnt_only_plugins = [straxen.nveto_recorder,
-                    straxen.nveto_pulse_processing,
-                    straxen.nveto_hitlets,
-                    straxen.acqmon_processing,
-                    straxen.pulse_processing,
-                    straxen.peaklet_processing,
-                    straxen.peak_processing,
-                    ]
+# Plugins in these files have nT plugins, E.g. in pulse&peak(let)
+# processing there are plugins for High Energy plugins. Therefore do not
+# st.register_all in 1T contexts.
+have_nT_plugins = [straxen.nveto_recorder,
+                   straxen.nveto_pulse_processing,
+                   straxen.nveto_hitlets,
+                   straxen.acqmon_processing,
+                   straxen.pulse_processing,
+                   straxen.peaklet_processing,
+                   straxen.peak_processing,
+                   straxen.online_monitor,
+                   ]
 
 ##
 # XENONnT
@@ -60,9 +67,8 @@ xnt_only_plugins = [straxen.nveto_recorder,
 
 def xenonnt_online(output_folder='./strax_data',
                    we_are_the_daq=False,
-                   _minimum_run_number=9271,
+                   _minimum_run_number=7157,
                    _database_init=True,
-
                    **kwargs):
     """XENONnT online processing and analysis"""
     context_options = {
@@ -72,7 +78,7 @@ def xenonnt_online(output_folder='./strax_data',
     st = strax.Context(
         config=straxen.contexts.xnt_common_config,
         **context_options)
-    st.register_all(xnt_only_plugins)
+    st.register_all(have_nT_plugins)
     st.register([straxen.DAQReader, straxen.LEDCalibration])
 
     st.storage = [straxen.RunDB(
@@ -100,26 +106,15 @@ def xenonnt_online(output_folder='./strax_data',
     elif _database_init:
         st.storage += [straxen.OnlineMonitor(
             readonly=not we_are_the_daq,
-            take_only=('pulse_counts', 'pulse_counts_he', 'veto_intervals'))]
+            take_only=('veto_intervals',
+                       'online_monitor',))]
 
     # Remap the data if it is before channel swap (because of wrongly cabled
-    # signal cable connectors) These are runs older than run 8797, before
-    # commissioning. Runs newer than 8796 are not affected. See:
+    # signal cable connectors) These are runs older than run 8797. Runs
+    # newer than 8796 are not affected. See:
     # https://github.com/XENONnT/straxen/pull/166 and
     # https://xe1t-wiki.lngs.infn.it/doku.php?id=xenon:xenonnt:dsg:daq:sector_swap
     st.set_context_config({'apply_data_function': (straxen.common.remap_old,)})
-    return st
-
-
-def xenonnt_initial_commissioning(**kwargs):
-    """
-    First phase of the commissioning of XENONnT.
-    These are runs 7157-9271.
-    xe1t-wiki.lngs.infn.it/doku.php?id=xenon:xenonnt:analysis:commissioning:straxen_contexts
-    """
-    st = xenonnt_online(_minimum_run_number=7157, **kwargs)
-    st.set_config(dict(
-        gain_model=('to_pe_constant', 'TemporaryGXe_1500V_PMT116_1300_PMT195_1300')))
     return st
 
 
@@ -144,13 +139,43 @@ def xenonnt_simulation(output_folder='./strax_data'):
     st = strax.Context(
         storage=strax.DataDirectory(output_folder),
         config=dict(detector='XENONnT',
-                    fax_config=straxen.aux_repo + '4e71b8a2446af772c83a8600adc77c0c3b7e54d1/fax_files/fax_config_nt.json',
+                    fax_config=straxen.aux_repo + '17e83b1a8c02f56066081ebaefe5f625ebb2e287/fax_files/fax_config_nt.json',
+                    check_raw_record_overlaps=False,
                     **straxen.contexts.xnt_common_config,
                     ),
         **straxen.contexts.common_opts)
     st.register(wfsim.RawRecordsFromFaxNT)
     return st
 
+
+def xenonnt_temporary_five_pmts(**kwargs):
+    """Temporary context for selected PMTs"""
+    # Start from the online context
+    st_online = xenonnt_online(**kwargs)
+
+    temporary_five_pmts_config = {
+        'gain_model': ('CMT_model', ("to_pe_model", "xenonnt_temporary_five_pmts")),
+        'peak_min_pmts': 2,
+        'peaklet_gap_threshold': 300,
+    }
+    # If there are any config overwrites in the kwargs, us those,
+    # otherwise use the config as in the dict above.
+
+    for k in list(temporary_five_pmts_config.keys()):
+        if k in kwargs:
+            temporary_five_pmts_config[k] = kwargs[k]
+
+    # Copy the online context and change the configuration here
+    st = st_online.new_context()
+    st.set_config(temporary_five_pmts_config)
+
+    return st
+
+
+def xenonnt_initial_commissioning(*args, **kwargs):
+    raise ValueError(
+        'Use xenonnt_online. See' 
+        'https://xe1t-wiki.lngs.infn.it/doku.php?id=xenon:xenonnt:analysis:commissioning:straxen_contexts#update_09_nov_20')
 
 ##
 # XENON1T
@@ -282,6 +307,7 @@ def xenon1t_simulation(output_folder='./strax_data'):
         config=dict(
             fax_config=straxen.aux_repo + '1c5793b7d6c1fdb7f99a67926ee3c16dd3aa944f/fax_files/fax_config_1t.json',
             detector='XENON1T',
+            check_raw_record_overlaps=False,
             **straxen.contexts.x1t_common_config),
         **straxen.contexts.common_opts)
     st.register(wfsim.RawRecordsFromFax1T)