+++ straxen/mini_analysis.py
@@ -73,7 +73,7 @@ def mini_analysis(requires=tuple(),
             kwargs.setdefault('time_selection', default_time_selection)
             kwargs.setdefault('selection_str', None)
 
-            kwargs['t_reference'] = context.estimate_run_start(
+            kwargs['t_reference'], _ = context.estimate_run_start_and_end(
                 run_id, requires)
 
             if warn_beyond_sec is not None and not kwargs.get('ignore_time_warning'):
@@ -112,7 +112,9 @@ def mini_analysis(requires=tuple(),
                             # Arguments for new context, if needed
                             config=kwargs.get('config'),
                             register=kwargs.get('register'),
-                            storage=kwargs.get('storage', tuple()))
+                            storage=kwargs.get('storage', tuple()),
+                            progress_bar=False,
+                        )
 
                 # If user did not give time kwargs, but the function expects
                 # a time_range, try to add one based on the time range of the data